# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Ascent 1.0 - Operational Intelligence Platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      summary: List all workflows
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, paused, at_risk]
        - name: stoplight
          in: query
          schema:
            type: string
            enum: [red, yellow, green]
      responses:
        "200":
          description: List of workflows
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Workflow"
    post:
      operationId: createWorkflow
      tags: [workflows]
      summary: Create a new workflow
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkflowBody"
      responses:
        "201":
          description: Created workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
  /workflows/{id}:
    get:
      operationId: getWorkflow
      tags: [workflows]
      summary: Get workflow by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowDetail"
    put:
      operationId: updateWorkflow
      tags: [workflows]
      summary: Update a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateWorkflowBody"
      responses:
        "200":
          description: Updated workflow
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workflow"
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      summary: Delete a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /workflows/{id}/stages:
    get:
      operationId: listStages
      tags: [stages]
      summary: List stages for a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Stage"
    post:
      operationId: createStage
      tags: [stages]
      summary: Create a stage for a workflow
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStageBody"
      responses:
        "201":
          description: Created stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
  /workflows/{id}/stages/{stageId}:
    put:
      operationId: updateStage
      tags: [stages]
      summary: Update a workflow stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStageBody"
      responses:
        "200":
          description: Updated stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stage"
    delete:
      operationId: deleteStage
      tags: [stages]
      summary: Delete a workflow stage
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: stageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Stoplight:
      type: string
      enum: [red, yellow, green]
...
    Workflow:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, paused, at_risk]
        stoplight:
          $ref: "#/components/schemas/Stoplight"
        flowScore:
          type: number
        riskScore:
          type: number
        improvementScore:
          type: number
        executionScore:
          type: number
        healthScore:
          type: number
        owner:
          type: string
          nullable: true
        stagesCount:
          type: integer
        completedStagesCount:
          type: integer
        dueDate:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - title
        - status
        - stoplight
        - flowScore
        - riskScore
        - improvementScore
        - executionScore
        - healthScore
        - stagesCount
        - completedStagesCount
        - createdAt
        - updatedAt
    WorkflowDetail:
      allOf:
        - $ref: "#/components/schemas/Workflow"
        - type: object
          properties:
            stages:
              type: array
              items:
                $ref: "#/components/schemas/Stage"
            documents:
              type: array
              items:
                $ref: "#/components/schemas/Document"
            alerts:
              type: array
              items:
                $ref: "#/components/schemas/Alert"
            impactEvents:
              type: array
              items:
                $ref: "#/components/schemas/ImpactEvent"
          required: [stages, documents, alerts, impactEvents]
...
    CreateWorkflowBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
      required: [title]
    UpdateWorkflowBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, completed, paused, at_risk]
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
    Stage:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        name:
          type: string
        order:
          type: integer
        status:
          type: string
          enum: [pending, in_progress, completed, blocked, overdue]
        stoplight:
          $ref: "#/components/schemas/Stoplight"
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        startedAt:
          type: string
          nullable: true
        completedAt:
          type: string
          nullable: true
        isBottleneck:
          type: boolean
        daysInStage:
          type: number
          nullable: true
        notes:
          type: string
          nullable: true
      required:
        - id
        - workflowId
        - name
        - order
        - status
        - stoplight
        - isBottleneck
    CreateStageBody:
      type: object
      properties:
        name:
          type: string
        order:
          type: integer
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, order]
    UpdateStageBody:
      type: object
      properties:
        name:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, blocked, overdue]
        owner:
          type: string
          nullable: true
        dueDate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        isBottleneck:
          type: boolean
...
    Document:
      type: object
      properties:
        id:
          type: integer
        filename:
          type: string
        url:
          type: string
          nullable: true
        detectedType:
          type: string
          nullable: true
        workflowId:
          type: integer
          nullable: true
        stageId:
          type: integer
          nullable: true
        assetId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, filename, createdAt]
...
    Alert:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [status_change, due_date, bottleneck, warranty_expiration, risk_escalation, overdue]
        severity:
          type: string
          enum: [critical, warning, info]
        title:
          type: string
        message:
          type: string
        workflowId:
          type: integer
          nullable: true
        assetId:
          type: integer
          nullable: true
        isRead:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - type
        - severity
        - title
        - message
        - isRead
        - createdAt
    ImpactEvent:
      type: object
      properties:
        id:
          type: integer
        workflowId:
          type: integer
        eventType:
          type: string
          enum: [delay, approval, change_order, risk_escalation, completion]
        description:
          type: string
        costImpact:
          type: number
          nullable: true
        timeImpactDays:
          type: number
          nullable: true
        createdAt:
          type: string
      required: [id, workflowId, eventType, description, createdAt]
...