# lib/api-zod/src/generated/types/updateWorkflowBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ascent 1.0 - Operational Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateWorkflowBodyStatus } from "./updateWorkflowBodyStatus";

export interface UpdateWorkflowBody {
  title?: string;
  description?: string | null;
  status?: UpdateWorkflowBodyStatus;
  owner?: string | null;
  dueDate?: string | null;
}