# lib/api-zod/src/generated/types/workflow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ascent 1.0 - Operational Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Stoplight } from "./stoplight";
import type { WorkflowStatus } from "./workflowStatus";

export interface Workflow {
  id: number;
  title: string;
  description?: string | null;
  status: WorkflowStatus;
  stoplight: Stoplight;
  flowScore: number;
  riskScore: number;
  improvementScore: number;
  executionScore: number;
  healthScore: number;
  owner?: string | null;
  stagesCount: number;
  completedStagesCount: number;
  dueDate?: string | null;
  createdAt: string;
  updatedAt: string;
}