# lib/api-zod/src/generated/types/workflowDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Ascent 1.0 - Operational Intelligence Platform
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";
import type { Document } from "./document";
import type { ImpactEvent } from "./impactEvent";
import type { Stage } from "./stage";
import type { Workflow } from "./workflow";

export type WorkflowDetail = Workflow & {
  stages: Stage[];
  documents: Document[];
  alerts: Alert[];
  impactEvents: ImpactEvent[];
};